Chore/implement testing
* Implemented checks to validate user creation

+++ src/main/java/com/itesm/application/dto/UserProfileDto.java
@@ -16,7 +16,7 @@ public class UserProfileDto {
     @JsonProperty("last_name2")
     private String lastName2;
 
-    private Integer age;
+    private Byte age;
 
     private SuburbDto suburb;
